shell: move window manipulation events to trx
Resolves #3418.
