+++ src/libtrx/game/shell/events.c
@@ -11,6 +11,15 @@ static bool m_ConsoleJustOpened = false;
 static void M_HandleQuit(void);
 static void M_HandleKeyDown(const SDL_Event *event);
 static void M_HandleKeyUp(const SDL_Event *event);
+static void M_HandleFocusGained(void);
+static void M_HandleFocusLost(void);
+static void M_HandleWindowShown(void);
+static void M_HandleWindowRestored(void);
+static void M_HandleWindowMinimized(void);
+static void M_HandleWindowMaximized(void);
+static void M_HandleWindowMoved(int32_t x, int32_t y);
+static void M_HandleWindowResized(int32_t width, int32_t height);
+static bool M_CreateGameWindow(void);
 
 static void M_HandleQuit(void)
 {
@@ -43,6 +52,44 @@ static void M_HandleKeyUp(const SDL_Event *const event)
     if (event->key.keysym.sym == SDLK_PRINTSCREEN) {
         Screenshot_Make(g_Config.rendering.screenshot_format);
     }
+}
+
+static void M_HandleFocusGained(void)
+{
+}
+
+static void M_HandleFocusLost(void)
+{
+}
+
+static void M_HandleWindowShown(void)
+{
+    LOG_DEBUG("");
+}
+
+static void M_HandleWindowRestored(void)
+{
+    Shell_SyncFromWindow(true);
+}
+
+static void M_HandleWindowMinimized(void)
+{
+    LOG_DEBUG("");
+}
+
+static void M_HandleWindowMaximized(void)
+{
+    Shell_SyncFromWindow(true);
+}
+
+static void M_HandleWindowMoved(const int32_t x, const int32_t y)
+{
+    Shell_SyncFromWindow(false);
+}
+
+static void M_HandleWindowResized(int32_t width, int32_t height)
+{
+    Shell_SyncFromWindow(true);
 }
 
 bool Shell_ProcessCommonEvent(const SDL_Event *const event)
@@ -82,6 +129,42 @@ bool Shell_ProcessCommonEvent(const SDL_Event *const event)
     case SDL_JOYDEVICEREMOVED:
         Input_Discover();
         return true;
+
+    case SDL_WINDOWEVENT:
+        switch (event->window.event) {
+        case SDL_WINDOWEVENT_SHOWN:
+            M_HandleWindowShown();
+            break;
+
+        case SDL_WINDOWEVENT_FOCUS_GAINED:
+            M_HandleFocusGained();
+            break;
+
+        case SDL_WINDOWEVENT_FOCUS_LOST:
+            M_HandleFocusLost();
+            break;
+
+        case SDL_WINDOWEVENT_RESTORED:
+            M_HandleWindowRestored();
+            break;
+
+        case SDL_WINDOWEVENT_MINIMIZED:
+            M_HandleWindowMinimized();
+            break;
+
+        case SDL_WINDOWEVENT_MAXIMIZED:
+            M_HandleWindowMaximized();
+            break;
+
+        case SDL_WINDOWEVENT_MOVED:
+            M_HandleWindowMoved(event->window.data1, event->window.data2);
+            break;
+
+        case SDL_WINDOWEVENT_RESIZED:
+            M_HandleWindowResized(event->window.data1, event->window.data2);
+            break;
+        }
+        break;
     }
 
     return false;
