+++ src/libtrx/meson.build
@@ -248,6 +248,7 @@ sources = [
   'game/savegame/common.c',
   'game/scaler.c',
   'game/shell/common.c',
+  'game/shell/config.c',
   'game/shell/events.c',
   'game/shell/flow.c',
   'game/shell/main.c',
