+++ src/tr1/game/output.h
@@ -8,7 +8,6 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void Output_SetWindowSize(int32_t width, int32_t height);
 void Output_ApplyLevelSettings(void);
 void Output_ApplyRenderSettings(void);
 void Output_ObserveFOVChange(void);
